refactor contract and types
