+++ apps/contract/admin/schema.ts
@@ -1,5 +1,6 @@
 import { z } from "zod";
+import { EmailSchema } from "../common";
 
 export const verifyAdminRequestSchema = z.object({
-  email: z.string(),
+  email: EmailSchema,
 });
