--- apps/contract/adminAuth/schema.ts
@@ -1,7 +1,0 @@
-import { z } from "zod";
-import { UserBaseSchema } from "../user/schema";
-
-export const CreateAdminSchema = UserBaseSchema.omit({
-  verficationId: true,
-  identityCardUrl: true,
-});
