+++ apps/contract/index.ts
@@ -6,11 +6,13 @@ import { uploadContract } from "./upload/contract";
 import { accountContract } from "./account/contract";
 import { genreContract } from "./genre/contract";
 import { bookContract } from "./book/contract";
+import { adminContract } from "./admin/contract";
 
 const c = initContract();
 
 export const contract = c.router({
   adminAuth: adminAuthContract,
+  admin: adminContract,
   userAuth: userAuthContract,
   user: userContract,
   upload: uploadContract,
