+++ apps/contract/user/contract.ts
@@ -1,8 +1,6 @@
+import { z } from "zod";
 import { initContract } from "@ts-rest/core";
 import {
-  UserBaseSchema,
-  CreateUserSuccessSchema,
-  UserDetailsSchema,
   GetAllUserDetailsQuerySchema,
   GetAllUserSchema,
   ChangeStatusRequestSchema,
@@ -10,8 +8,13 @@ import {
   GetAllAccountQuerySchema,
   GetAllAccountRequestSchema,
 } from "../user/schema";
-import { PaginatedRequestSchema, SuccessSchema } from "../common";
-import { z } from "zod";
+import {
+  PaginatedRequestSchema,
+  SuccessSchema,
+  CreateUserSuccessSchema,
+  UserBaseSchema,
+  UserDetailsSchema,
+} from "../common";
 
 const c = initContract();
